\name{swemap-package}
\alias{swemap-package}
\alias{swemap}
\docType{package}
\title{
\packageTitle{swemap}
}
\description{
\packageDescription{swemap}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{swemap}
\packageIndices{swemap}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{swemap}

Maintainer: \packageMaintainer{swemap}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation ~~
~~ directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
